pretty printing works or unix

+++ src/cli_remove/remove_tls_secret.rs
@@ -1,28 +1,37 @@
 use crate::shared::file_folder_paths::{get_binary_path, Binary};
-use crate::shared::{get_k8s_client, handle_output, TLS_SECRET};
+use crate::shared::{
+    get_k8s_client, handle_output, restore_term, save_term, should_continue_as_admin, TLS_SECRET,
+};
 use anyhow::{anyhow, Result};
+
 use k8s_openapi::api::core::v1::Secret;
 use kube::api::DeleteParams;
 use kube::{Api, Error};
 use std::process::Command;
 
 pub async fn remove_tls_secret() -> Result<()> {
+    if !should_continue_as_admin()? {
+        return Err(anyhow!("skipped"));
+    }
     remove_local_ca()?;
     delete_tls_secret().await?;
 
     Ok(())
 }
 
-// todo: admin warning
 fn remove_local_ca() -> Result<()> {
     let mkcert_path = get_binary_path(Binary::Mkcert)?;
 
     if mkcert_path.exists() {
-        let output = Command::new(mkcert_path)
+        save_term()?;
+
+        let output = Command::new("sudo")
+            .arg(mkcert_path)
             .arg("-uninstall")
             .output()
             .expect("failed to execute process");
 
+        restore_term(1)?;
         handle_output(output)?;
     }
